 \name{Deccan}
\alias{Deccan}
\docType{data}
\title{Data collection of igneous rocks from the Deccan region (India)}
\description{
Compilation of major and trace element analysis of 7019 igneous rocks from the Deccan region, India.
}
\usage{data(Deccan)}
\format{
The format is a data frame with 7019 rows and 30 columns.
}
\details{The database is part of a global whole-rock geochemical database, sourced from various existing databases and supplemented with an extensive list of individual publications.
}

\source{
The database is part of the compiled data in Gard et al. (2019). For full details see \url{https://doi.org/10.5194/essd-11-1553-2019} . 
}

\references{
Gard, M., Hasterok, D. and Halpin, J.A., 2019. Global whole-rock geochemical database compilation. Earth System Science Data, 11(4), pp.1553-1566.
}
\examples{
data(Deccan)
##
}
\seealso{
\code{\link{EAP}}
\code{\link{IUGS}}
\code{\link{TephraKam}}
}
\keyword{datasets}


