\name{IUGS}
\alias{IUGS}
\docType{data}
\title{Data collection of igneous rocks used in IUGSTAS software}
\description{
Compilation of major elements data for 37 representative samples of volcanic rocks.
}
\usage{data(TestTAS)}
\format{
  The format is a data frame with 37 rows and 28 columns.
}
\details{
This data is available from the Cambridge University Press, as a part of IUGSTAS software. 
}

\references{
Le Maitre, R.W et al. 2002. Igneous Rocks: A Classification and Glossary of Terms: Recommendations of the International Union of Geological Sciences, Subcommission on the Systematics of Igneous Rocks. Cambridge University Press, 2002, 252 pp.
}
\examples{
data(IUGS)
##
}

\seealso{
\code{\link{Deccan}}
\code{\link{EAP}}
\code{\link{TephraKam}}
}

\keyword{datasets}


