\docType{package}
\name{shinyNORRRM-package}
\alias{shinyNORRRM-app}
\alias{shinyNORRRM-package}
\title{shinyNORRRM-: The ultimate CIPW Norm}
\description{
\if{html}{\figure{logos/logo.png}{options: style='float: right' alt='logo' width='120'}}

The computer program is an efficient CIPW algorithm and igneous rock classification system written in R and running in shiny.
}
\details{
```
install.packages("shinyNORRRM")
shinyNORRRM::shinyNORRRM()
```

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TheRFrog/shinyNORRRM}
  \item Report bugs at \url{https://github.com/TheRFrog/shinyNORRRM/issues}
}
}
\author{
\strong{Maintainer}: Reneé González-Guzmán \email{rguzman@geociencias.unam.mx} (\href{https://orcid.org/0000-0001-7773-6401}{ORCID})
}

\keyword{internal}
