 \name{TephraKam}
\alias{TephraKam}
\docType{data}
\title{Data collection of igneous rocks from the Kamchatka volcanic arc (northwestern Pacific)}
\description{
Compilation of major and minor element analysis of 7596 igneous rocks from Kamchatka, Rusia.
}
\usage{data(TephraKam)}
\format{
The format is a data frame with 7596 rows and 30 columns.
}
\details{The database contains 7596 single-shard major and minor element analyses obtained by electron microprobe. The samples characterize about 300 explosive eruptions, which occurred in Kamchatka from the Miocene up to recent times.
}

\source{
The database is also available in Portnyagin et al. (2019). For full details see \url{https://doi.org/10.5194/essd-12-469-2020} . 
}

\references{
Portnyagin, M.V., Ponomareva, V.V., Zelenin, E.A., Bazanova, L.I., Pevzner, M.M., Plechova, A.A., Rogozin, A.N. and Garbe-Schönberg, D., 2020. TephraKam: geochemical database of glass compositions in tephra and welded tuffs from the Kamchatka volcanic arc (northwestern Pacific). Earth System Science Data, 12(1), pp.469-486.
}
\examples{
data(TephraKam)
##
}
\seealso{
\code{\link{Deccan}}
\code{\link{EAP}}
\code{\link{IUGS}}
}
\keyword{datasets}


